refactor: Move config.dart into core dir
